Code: Add rival starters basic functionality
What this does is basically storing the rivals starter
mon that is made based on player's starter.

Now, I know this might be hacky but tbh I kinda want
to do wacky shit like this- For educational purpose
to myself, of course.

Just need to do extra mods for when to evolve their starter and also put it into their trainer party in battle.